Move static registry to new file
